SNOW-206907 GCS Downscope Token
Description
- Implement update to work with downscoped GoogleCredential instead of using
presigned url. To use this new feature, SNOW-206907 must be reviewed and merged
on the server side first.

Testing
- Add test to use GoogleCredential instead of presigned URL by setting a parameter.
- Overwrite false digest test is also changed to reflect this. With 206907,
OVERWRITE parameter can be used with GCS PUT command.
