+++ src/main/java/net/snowflake/client/jdbc/SnowflakeFileTransferAgent.java
@@ -2204,9 +2204,12 @@ public class SnowflakeFileTransferAgent implements SnowflakeFixedView {
         } catch (Exception ex) {
           logger.debug("Listing objects for filtering encountered exception: {}", ex.getMessage());
 
-          // Need to unwrap StorageProviderException since handleStorageException only handle base cause.
+          // Need to unwrap StorageProviderException since handleStorageException only handle base
+          // cause.
           if (ex instanceof StorageProviderException) {
-            ex = (Exception) ex.getCause(); // Cause of StorageProviderException is always an Exception
+            ex =
+                (Exception)
+                    ex.getCause(); // Cause of StorageProviderException is always an Exception
           }
           storageClient.handleStorageException(ex, ++retryCount, "listObjects", session, command);
         }
